Move remaining org/project pages to projectAdministration
