+++ src/modules/projectAdministration/components/CreateProjectPage.tsx
@@ -4,10 +4,10 @@ import { useNavigate } from 'react-router-dom';
 
 import NotFound from '../../../components/pages/NotFound';
 
+import { useOrganizationCrumbs } from '../hooks/useOrganizationCrumbs';
 import BasicBreadcrumbPageLayout from '@/components/layout/BasicBreadcrumbPageLayout';
 import useSafeParams from '@/hooks/useSafeParams';
 import EditRecord from '@/modules/form/components/EditRecord';
-import { useOrganizationCrumbs } from '@/modules/projects/hooks/useOrganizationCrumbs';
 import { cache } from '@/providers/apolloClient';
 import { Routes } from '@/routes/routes';
 import { RecordFormRole, ProjectAllFieldsFragment } from '@/types/gqlTypes';
