+++ src/modules/projectAdministration/components/EditOrganizationPage.tsx
@@ -2,15 +2,15 @@ import { Stack, Typography } from '@mui/material';
 import { useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import Loading from '../elements/Loading';
+import Loading from '../../../components/elements/Loading';
 
-import NotFound from './NotFound';
+import NotFound from '../../../components/pages/NotFound';
 
+import { useOrganizationCrumbs } from '../hooks/useOrganizationCrumbs';
 import BasicBreadcrumbPageLayout from '@/components/layout/BasicBreadcrumbPageLayout';
 import DeleteMutationButton from '@/modules/dataFetching/components/DeleteMutationButton';
 import EditRecord from '@/modules/form/components/EditRecord';
 import { OrganizationPermissionsFilter } from '@/modules/permissions/PermissionsFilters';
-import { useOrganizationCrumbs } from '@/modules/projects/hooks/useOrganizationCrumbs';
 import { Routes } from '@/routes/routes';
 import {
   DeleteOrganizationDocument,
@@ -23,7 +23,7 @@ import {
 import { evictPickList, evictQuery } from '@/utils/cacheUtil';
 import { generateSafePath } from '@/utils/pathEncoding';
 
-const EditOrganization = () => {
+const EditOrganizationPage = () => {
   const navigate = useNavigate();
 
   const { crumbs, loading, organization, organizationName } =
@@ -96,4 +96,4 @@ const EditOrganization = () => {
     </BasicBreadcrumbPageLayout>
   );
 };
-export default EditOrganization;
+export default EditOrganizationPage;
